build: Initialize go module
Start transition from Dep to go modules.

Dep was an official experiment to
implement a package manager for Go. As
of 2020, Dep is deprecated and archived
in favor of Go modules, which have had
official support since Go 1.11.

This commit is the first step and is
essentially:
go mod init github.com/nlnwa/veidemann-ooshandler